# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Maarkbh.API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/vehicle-models/search": {
      "post": {
        "tags": [
          "VehicleModels"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleModelSearchRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleModelSearchRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleModelSearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
    "/api/vehicles/{id}": {
      "get": {
        "tags": [
          "Vehicles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Vehicles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
    "/api/vehicles": {
      "post": {
        "tags": [
          "Vehicles"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "CleanlinessStatus": {
        "enum": [
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "ConditionGrade": {
        "enum": [
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "FuelLevel": {
        "enum": [
          0,
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "InsurancePricingRequest": {
        "type": "object",
        "properties": {
          "branchId": {
            "type": "integer",
            "format": "int64"
          },
          "insuranceCompanyId": {
            "type": "integer",
            "format": "int64"
          },
          "insuranceTypeId": {
            "type": "integer",
            "format": "int64"
          },
          "insurancePolicyNumber": {
            "type": "string",
            "nullable": true
          },
          "insuranceExpiryDate": {
            "type": "string",
            "format": "date"
          },
          "insuranceAmount": {
            "type": "number",
            "format": "double"
          },
          "dailyRate": {
            "type": "number",
            "format": "double"
          },
          "extraKilometerRate": {
            "type": "number",
            "format": "double"
          },
          "fullFuelRate": {
            "type": "number",
            "format": "double"
          },
          "lateHourRate": {
            "type": "number",
            "format": "double"
          },
          "isKilometerLimitEnabled": {
            "type": "boolean"
          },
          "dailyKilometerLimit": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PlateDocsRequest": {
        "type": "object",
        "properties": {
          "plateTypeId": {
            "type": "integer",
            "format": "int64"
          },
          "plateNumber": {
            "type": "string",
            "nullable": true
          },
          "plateFirstLetter": {
            "type": "string",
            "nullable": true
          },
          "plateSecondLetter": {
            "type": "string",
            "nullable": true
          },
          "plateThirdLetter": {
            "type": "string",
            "nullable": true
          },
          "registrationNumber": {
            "type": "string",
            "nullable": true
          },
          "registrationExpiryDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "inspectionExpiryDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "serialNumber": {
            "type": "string",
            "nullable": true
          },
          "operationCardNumber": {
            "type": "string",
            "nullable": true
          },
          "operationCardExpiryDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "customsNumber": {
            "type": "string",
            "nullable": true
          },
          "otherNotes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PresenceStatus": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "TajeerStatusRequest": {
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/VehicleFleetStatus"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "isListingActive": {
            "type": "boolean"
          },
          "odometerReading": {
            "type": "integer",
            "format": "int32"
          },
          "fuelLevel": {
            "$ref": "#/components/schemas/FuelLevel"
          },
          "enduranceAmount": {
            "type": "number",
            "format": "double"
          },
          "oilType": {
            "$ref": "#/components/schemas/VehicleOilType"
          },
          "lastOilChangeDate": {
            "type": "string",
            "format": "date"
          },
          "oilChangeDistance": {
            "type": "integer",
            "format": "int32"
          },
          "airConditionGrade": {
            "$ref": "#/components/schemas/ConditionGrade"
          },
          "radioStatus": {
            "$ref": "#/components/schemas/WorkingStatus"
          },
          "screenStatus": {
            "$ref": "#/components/schemas/WorkingStatus"
          },
          "odometerStatus": {
            "$ref": "#/components/schemas/WorkingStatus"
          },
          "seatCleanliness": {
            "$ref": "#/components/schemas/CleanlinessStatus"
          },
          "keyStatus": {
            "$ref": "#/components/schemas/WorkingStatus"
          },
          "tireCondition": {
            "$ref": "#/components/schemas/TireCondition"
          },
          "spareTireStatus": {
            "$ref": "#/components/schemas/PresenceStatus"
          },
          "fireExtinguisherStatus": {
            "$ref": "#/components/schemas/PresenceStatus"
          },
          "firstAidKitStatus": {
            "$ref": "#/components/schemas/PresenceStatus"
          },
          "safetyTriangleStatus": {
            "$ref": "#/components/schemas/PresenceStatus"
          },
          "tireToolsStatus": {
            "$ref": "#/components/schemas/PresenceStatus"
          }
        },
        "additionalProperties": false
      },
      "TireCondition": {
        "enum": [
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "VehicleBodyType": {
        "enum": [
          1,
          2,
          3,
          4,
          5,
          6,
          7
        ],
        "type": "integer",
        "format": "int32"
      },
      "VehicleCategory": {
        "enum": [
          1,
          2,
          3,
          4,
          5,
          6,
          7
        ],
        "type": "integer",
        "format": "int32"
      },
      "VehicleDamagePointDto": {
        "type": "object",
        "properties": {
          "damageType": {
            "$ref": "#/components/schemas/VehicleDamageType"
          },
          "positionX": {
            "type": "number",
            "format": "double"
          },
          "positionY": {
            "type": "number",
            "format": "double"
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "fileId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "VehicleDamageType": {
        "enum": [
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32"
      },
      "VehicleFleetStatus": {
        "enum": [
          1,
          2,
          3,
          4,
          5,
          6,
          7
        ],
        "type": "integer",
        "format": "int32"
      },
      "VehicleFuelType": {
        "enum": [
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32"
      },
      "VehicleImageDto": {
        "type": "object",
        "properties": {
          "slotCode": {
            "type": "string",
            "nullable": true
          },
          "fileId": {
            "type": "integer",
            "format": "int64"
          },
          "isPrimary": {
            "type": "boolean"
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "VehicleInfoRequest": {
        "type": "object",
        "properties": {
          "makeId": {
            "type": "integer",
            "format": "int64"
          },
          "modelId": {
            "type": "integer",
            "format": "int64"
          },
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "color": {
            "type": "string",
            "nullable": true
          },
          "vin": {
            "type": "string",
            "nullable": true
          },
          "engineNumber": {
            "type": "string",
            "nullable": true
          },
          "bodyType": {
            "$ref": "#/components/schemas/VehicleBodyType"
          },
          "category": {
            "$ref": "#/components/schemas/VehicleCategory"
          },
          "seats": {
            "type": "integer",
            "format": "int32"
          },
          "cylinders": {
            "type": "integer",
            "format": "int32"
          },
          "fuelType": {
            "$ref": "#/components/schemas/VehicleFuelType"
          },
          "transmissionType": {
            "$ref": "#/components/schemas/VehicleTransmissionType"
          },
          "payloadKg": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "VehicleModelSearchRequest": {
        "type": "object",
        "properties": {
          "search": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean",
            "nullable": true
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "parentId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "VehicleOilType": {
        "enum": [
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "VehicleRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "plate": {
            "$ref": "#/components/schemas/PlateDocsRequest"
          },
          "info": {
            "$ref": "#/components/schemas/VehicleInfoRequest"
          },
          "insurancePricing": {
            "$ref": "#/components/schemas/InsurancePricingRequest"
          },
          "tajeerStatus": {
            "$ref": "#/components/schemas/TajeerStatusRequest"
          },
          "featureTypeIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "nullable": true
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VehicleImageDto"
            },
            "nullable": true
          },
          "damagePoints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VehicleDamagePointDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "VehicleTransmissionType": {
        "enum": [
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "WorkingStatus": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter the JWT access token only. Swagger will add the Bearer prefix.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}